requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
matplotlib
numpy
pandas
joblib
requests
pillow